Move MockServer to a separate class
